% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAnim-package.R
\docType{package}
\name{dataAnim-package}
\alias{dataAnim}
\alias{dataAnim-package}
\title{dataAnim: Visualise data transformation with animations}
\description{
Many data wrangling operations can seem quite abstract,
    making them harder to understand for people new to data science.
    This package makes the ideas more concrete, and hence easier to learn, 
    by animating transformation of data tables.
}
\author{
\strong{Maintainer}: Charco Hui \email{phui285@aucklanduni.ac.nz}

}
\keyword{internal}
