% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{load_toydata}
\alias{load_toydata}
\alias{datoy1}
\alias{datoy_long}
\alias{datoy_wide}
\title{Toy data sets}
\format{An object of class \code{list} of length 2.}
\usage{
load_toydata(num)

datoy1

datoy_long

datoy_wide
}
\arguments{
\item{num}{ID of the toy data set.}
}
\description{
Load toy data sets shown in the examples.
}
\examples{
## To load toy data set 1
load_toydata(1)
}
\author{
Charco Hui
}
\keyword{datasets}
