% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_anim.R
\name{spread_anim}
\alias{spread_anim}
\title{Spread Animation}
\usage{
spread_anim(key, value, data, speed = 1, width = NULL, height = NULL)
}
\arguments{
\item{key}{Column used to spread out to multiple columns.}

\item{value}{Column containing values of the key.}

\item{data}{Data to pass into the function.}

\item{speed}{Speed of the animation.}

\item{width}{Width of the animation frame in pixels.}

\item{height}{Height of the animation frame in pixels.}
}
\description{
Long to Wide transformation.
}
\examples{
data(toyda_long)
spread_anim(key = "Subject", value = "Score", data = toyda_long)
myanim = spread_anim(key = "Subject", value = "Score", data = toyda_long)
htmlwidgets::saveWidget(myanim, file = "myanim.html")
}
\author{
Charco Hui
}
