% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{load_toydata}
\alias{load_toydata}
\title{Toy data sets}
\usage{
load_toydata(num)
}
\arguments{
\item{num}{ID of the toy data set.}
}
\description{
Load toy data sets shown in the examples.
}
\examples{
## To load toy data set 1
load_toydata(1)
}
\author{
Charco Hui
}
